#10 secure unusual use cases

+++ src/main/java/com/task/collectionboxesforfundraisingevents/service/collectionBox/impl/CollectionBoxServiceImpl.java
@@ -4,6 +4,7 @@ import com.task.collectionboxesforfundraisingevents.entity.CollectionBox;
 import com.task.collectionboxesforfundraisingevents.entity.CollectionBoxContent;
 import com.task.collectionboxesforfundraisingevents.entity.FundraisingEvent;
 import com.task.collectionboxesforfundraisingevents.entity.FundraisingEventAccount;
+import com.task.collectionboxesforfundraisingevents.model.Currency;
 import com.task.collectionboxesforfundraisingevents.repository.CollectionBoxContentRepository;
 import com.task.collectionboxesforfundraisingevents.repository.CollectionBoxRepository;
 import com.task.collectionboxesforfundraisingevents.repository.FundraisingEventAccountRepository;
@@ -18,6 +19,7 @@ import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
@@ -32,6 +34,8 @@ public class CollectionBoxServiceImpl implements CollectionBoxService {
     private final FundraisingEventAccountRepository fundraisingEventAccountRepository;
     private final ExchangeService exchangeService;
 
+    @Override
+    @Transactional
     public CollectionBoxDto createCollectionBox() {
         return new CollectionBoxDto(collectionBoxRepository.save(new CollectionBox()));
     }
@@ -42,7 +46,11 @@ public class CollectionBoxServiceImpl implements CollectionBoxService {
     }
 
     @Override
+    @Transactional
     public void removeCollectionBox(Integer id) {
+        if (!collectionBoxRepository.existsById(id))
+            throw new EntityNotFoundException("Collection Box with id " + id + " not found");
+
         collectionBoxRepository.deleteById(id);
     }
 
@@ -52,11 +60,11 @@ public class CollectionBoxServiceImpl implements CollectionBoxService {
         CollectionBox box = collectionBoxRepository.findById(collectionBoxId)
                 .orElseThrow(() -> new EntityNotFoundException("Collection box not found"));
 
-        if (!box.isEmpty())
-            throw new IllegalStateException("Collection box must be empty before assigning");
-
         FundraisingEvent event = fundraisingEventRepository.findById(fundraisingEventId)
                 .orElseThrow(() -> new EntityNotFoundException("Fundraising event not found"));
+
+        if (!box.isEmpty())
+            throw new IllegalStateException("Collection box must be empty before assigning");
 
         box.setFundraisingEvent(event);
         box.setAssigned(true);
@@ -69,14 +77,20 @@ public class CollectionBoxServiceImpl implements CollectionBoxService {
         CollectionBox box = collectionBoxRepository.findById(collectionBoxId)
                 .orElseThrow(() -> new EntityNotFoundException("Collection box not found"));
 
+        if (!Currency.isValid(currency))
+            throw new IllegalArgumentException(
+                    "Invalid currency: " + currency + ". Currency must be one of: " + Arrays.toString(Currency.values())
+            );
+
+        if(Objects.isNull(amount) || amount.compareTo(BigDecimal.ZERO) <= 0)
+            throw new IllegalArgumentException("Amount must be greater than zero");
+
         Optional<CollectionBoxContent> optionalContent = collectionBoxContentRepository.findByCollectionBoxIdAndCurrency(collectionBoxId, currency);
 
         CollectionBoxContent content = optionalContent.orElseGet(() -> {
             CollectionBoxContent newContent = new CollectionBoxContent();
             newContent.setCollectionBox(box);
             newContent.setCurrency(currency);
-            newContent.setAmount(BigDecimal.ZERO);
-            box.setIsEmpty(false);
             collectionBoxRepository.save(box);
             return newContent;
         });
@@ -91,6 +105,9 @@ public class CollectionBoxServiceImpl implements CollectionBoxService {
     public void transferMoneyToFundraisingEvent(Integer collectionBoxId) {
         CollectionBox box = collectionBoxRepository.findById(collectionBoxId)
                 .orElseThrow(() -> new EntityNotFoundException("Collection box not found"));
+
+        if (box.getIsEmpty())
+            throw new IllegalStateException("Collection box is empty");
 
         if (box.getFundraisingEvent() == null)
             throw new IllegalStateException("Collection box is not assigned to any fundraising event");
